ya casi esta el ORM

+++ src/main/java/Grupo11/DDS_TP_Integrador/Georef/GeorefService.java
@@ -7,6 +7,8 @@ import retrofit2.http.Query;
 
 
 public interface GeorefService {
+
+    //OJO! LOS TIPOS DE DATO LONG PODRIAN SER INT
     @GET("provincias")
     Call<ListadoDeProvincias> provincias();
 
@@ -14,11 +16,11 @@ public interface GeorefService {
     Call<ListadoDeProvincias> provincias(@Query("campos") String campos);
 
     @GET("municipios")
-    Call<ListadoDeMunicipios> municipios(@Query("provincia") int idProvincia);
+    Call<ListadoDeMunicipios> municipios(@Query("provincia") long idProvincia);
 
     @GET("municipios")
-    Call<ListadoDeMunicipios> municipios(@Query("provincia") int idProvincia, @Query("campos") String campos);
+    Call<ListadoDeMunicipios> municipios(@Query("provincia") long idProvincia, @Query("campos") String campos);
 
     @GET("municipios")
-    Call<ListadoDeMunicipios> municipios(@Query("provincia") int idProvincia, @Query("campos") String campos, @Query("max") int max);
+    Call<ListadoDeMunicipios> municipios(@Query("provincia") long idProvincia, @Query("campos") String campos, @Query("max") int max);
 }
